+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: arnduran <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/26 18:20:35 by arnduran          #+#    #+#             */
-/*   Updated: 2018/11/26 20:20:54 by arnduran         ###   ########.fr       */
+/*   Updated: 2019/01/23 19:36:02 by arnduran         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strncat(char *dest, const char *src, size_t n)
 {
-	int	i;
-	int j;
+	size_t	i;
+	size_t	j;
 
 	j = 0;
 	i = ft_strlen(dest);
